Admin permissions

+++ src/main/java/me/dionclei/workshopspringboot/resources/UserResource.java
@@ -2,6 +2,7 @@ package me.dionclei.workshopspringboot.resources;
 
 import java.lang.System.Logger;
 import java.net.URI;
+import java.security.Principal;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,6 +11,7 @@ import java.util.stream.Collectors;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -29,6 +31,7 @@ import me.dionclei.workshopspringboot.entities.dto.OrderItemRequest;
 import me.dionclei.workshopspringboot.entities.dto.OrderRequest;
 import me.dionclei.workshopspringboot.entities.dto.UserDTO;
 import me.dionclei.workshopspringboot.enums.OrderStatus;
+import me.dionclei.workshopspringboot.enums.UserRole;
 import me.dionclei.workshopspringboot.services.OrderService;
 import me.dionclei.workshopspringboot.services.ProductService;
 import me.dionclei.workshopspringboot.services.UserService;
@@ -57,8 +60,12 @@ public class UserResource {
 	}
 	
 	@GetMapping(value = "/{id}")
-	public ResponseEntity<User> findById(@PathVariable long id) {
-		return ResponseEntity.ok().body(userService.findById(id));
+	public ResponseEntity<UserDTO> findById(@PathVariable long id, Principal principal) {
+		var user = userService.findByEmail(principal.getName());
+		if(user.getId() != id && user.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) { 
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+	    return ResponseEntity.ok().body(userService.findById(id).toDTO());
 	}
 	
 	@PostMapping
@@ -69,25 +76,41 @@ public class UserResource {
 	}
 	
 	@DeleteMapping(value = "/{id}")
-	public ResponseEntity<Void> delete(@PathVariable Long id) {
+	public ResponseEntity<Void> delete(@PathVariable Long id, Principal principal) {
+		var user = userService.findByEmail(principal.getName());
+		if (id != user.getId() && user.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		userService.delete(id);
 		return ResponseEntity.noContent().build();
 	}
 	
 	@PutMapping(value = "/{id}")
-	public ResponseEntity<UserDTO> update(@PathVariable Long id, @RequestBody User obj) {
+	public ResponseEntity<UserDTO> update(@PathVariable Long id, @RequestBody User obj, Principal principal) {
+		var user = userService.findByEmail(principal.getName());
+		if (id != user.getId() && user.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		obj = userService.update(id, obj);
 		return ResponseEntity.ok().body(obj.toDTO());
 	}
 	
 	@GetMapping(value = "/{id}/orders")
-	public ResponseEntity<List<OrderDTO>> getOrders(@PathVariable Long id) {
+	public ResponseEntity<List<OrderDTO>> getOrders(@PathVariable Long id, Principal principal) {
+		var user = userService.findByEmail(principal.getName());
+		if (id != user.getId() && user.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		return ResponseEntity.ok(userService.getOrders(id)) ;
 	}
 	
 	@PutMapping(value = "/{id}/orders")
-	public ResponseEntity<OrderDTO> addOrder(@PathVariable Long id, @RequestBody OrderRequest orderRequest) {
-		User user = userService.findById(id);
+	public ResponseEntity<OrderDTO> addOrder(@PathVariable Long id, @RequestBody OrderRequest orderRequest, Principal principal) {
+		var user = userService.findByEmail(principal.getName());
+		if (id != user.getId() && user.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+		user = userService.findById(id);
 		Order order = new Order();
 		order.setClient(user);
 		order.setMoment(Instant.now());
